requirements.txt
sentence-transformers
langchain
flask
flask-cors
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_openai
langchain_community
langchain_experimental
-e .